modernize package structure
