+++ doc/changes.rst
@@ -2,10 +2,15 @@
 Change Log
 ==========
 
-1.2.2 (unreleased)
+2.0.0 (unreleased)
 ------------------
 
-* No changes yet.
+The plan is to drop Python 2 support in this version.
+
+1.3.0 (unreleased)
+------------------
+
+* Update Travis and ReadTheDocs configuration to use webhooks.
 
 1.2.1 (2017-10-26)
 ------------------
