+++ requirements.txt
@@ -1,1 +1,1 @@
-git+https://github.com/desihub/desiutil@1.9.3#egg=desiutil
+git+https://github.com/desihub/desiutil@1.9.13#egg=desiutil
