+++ setup.cfg
@@ -1,0 +1,13 @@
+[pycodestyle]
+# See https://pycodestyle.readthedocs.io/en/latest/intro.html#configuration
+# for details of these configuration options.
+#
+# These are normally ignored by default:
+# ignore = E121, E123, E126, E133, E226, E241, E242, E704, W503, W504
+#
+# In addition to the default set we add:
+# - E501: line too long (82 > 79 characters)
+# - E731: do not assign a lambda expression, use a def
+# - E741: do not use variables named 'l', 'O', or 'I' -- because, for example,
+#   'l' might refer to Galactic longitude.
+ignore = E121, E123, E126, E133, E226, E241, E242, E501, E704, E731, E741, W503, W504
